+++ packages/svelte/src/lib/components/EdgeLabel/EdgeLabel.svelte
@@ -19,6 +19,7 @@
     class="svelte-flow__edge-label"
     style:transform="translate(-50%, -50%) translate({x}px,{y}px)"
     style={'pointer-events: all;' + style}
+    role="button"
     on:click={() => {
       if (id) handleEdgeSelect(id);
     }}